Fix #4: use double for sortOrder, new snippets default to front of category
- Changed sortOrder from int to double for conflict-free ordering
- New snippets now get sortOrder = min(same_category) - 1.0
- Updated tests to verify descending sortOrder for new snippets

+++ src/snippet_manager.h
@@ -21,7 +21,7 @@ struct Snippet {
     QString compileCommand;
     QString code;
     bool isPreset = false;
-    int sortOrder = 0;
+    double sortOrder = 0.0;
 };
 Q_DECLARE_METATYPE(Snippet)
